// SysManager.UI.Admin/src/app/components/product/models/product-view.ts
export class ProductView{
    id: string='';
    name: string='';
    productCode: string = '';
    productTypeId: string = '';
    categoryId: string = '';
    unityId: string = '';
    costPrice: number = 0;
    percentage: number = 0;
    price: number = 0;
    active: boolean= false;
}